+++ trixnity-client/src/commonMain/kotlin/de/connect2x/trixnity/client/media/MediaService.kt
@@ -1,12 +1,10 @@
 package de.connect2x.trixnity.client.media
 
 import de.connect2x.lognity.api.logger.Logger
-import de.connect2x.trixnity.client.MatrixClientConfiguration
 import de.connect2x.trixnity.client.store.MediaCacheMapping
 import de.connect2x.trixnity.client.store.MediaCacheMappingStore
 import de.connect2x.trixnity.client.store.ServerDataStore
 import de.connect2x.trixnity.clientserverapi.client.MatrixClientServerApiClient
-import de.connect2x.trixnity.clientserverapi.client.FileSizeLimit
 import de.connect2x.trixnity.clientserverapi.client.MediaApiClient
 import de.connect2x.trixnity.clientserverapi.model.media.FileTransferProgress
 import de.connect2x.trixnity.clientserverapi.model.media.Media
@@ -37,14 +35,14 @@ interface MediaService {
         uri: String,
         progress: MutableStateFlow<FileTransferProgress?>? = null,
         saveToCache: Boolean = true,
-        maxSize: FileSizeLimit? = null,
+        maxSize: Long? = null,
     ): Result<PlatformMedia>
 
     suspend fun getEncryptedMedia(
         encryptedFile: EncryptedFile,
         progress: MutableStateFlow<FileTransferProgress?>? = null,
         saveToCache: Boolean = true,
-        maxSize: FileSizeLimit? = null,
+        maxSize: Long? = null,
     ): Result<PlatformMedia>
 
     suspend fun getThumbnail(
@@ -55,7 +53,7 @@ interface MediaService {
         animated: Boolean = false,
         progress: MutableStateFlow<FileTransferProgress?>? = null,
         saveToCache: Boolean = true,
-        maxSize: FileSizeLimit? = null,
+        maxSize: Long? = null,
     ): Result<PlatformMedia>
 
     suspend fun prepareUploadMedia(content: ByteArrayFlow, contentType: ContentType?): String
@@ -76,7 +74,6 @@ class MediaServiceImpl(
     private val mediaStore: MediaStore,
     private val serverDataStore: ServerDataStore,
     private val mediaCacheMappingStore: MediaCacheMappingStore,
-    private val config: MatrixClientConfiguration,
 ) : MediaService {
     companion object {
         private const val MATRIX_SPEC_1_11 = "v1.11"
@@ -97,7 +94,7 @@ class MediaServiceImpl(
     private suspend fun MediaApiClient.downloadDependingOnServerVersion(
         mxcUri: String,
         progress: MutableStateFlow<FileTransferProgress?>? = null,
-        maxSize: FileSizeLimit,
+        maxSize: Long?,
         downloadHandler: suspend (Media) -> Unit
     ): Result<Unit> =
         if (serverDataStore.getServerData().versions.versions.contains(MATRIX_SPEC_1_11)) {
@@ -112,7 +109,7 @@ class MediaServiceImpl(
         saveToCache: Boolean,
         sha256Hash: String?,
         progress: MutableStateFlow<FileTransferProgress?>?,
-        maxSize: FileSizeLimit,
+        maxSize: Long?,
     ): Result<PlatformMedia> = kotlin.runCatching {
         when {
             uri.startsWith(UPLOAD_MEDIA_MXC_URI_PREFIX) -> {
@@ -165,22 +162,20 @@ class MediaServiceImpl(
         uri: String,
         progress: MutableStateFlow<FileTransferProgress?>?,
         saveToCache: Boolean,
-        maxSize: FileSizeLimit?,
+        maxSize: Long?,
     ): Result<PlatformMedia> {
-        val maxSizeResolved = maxSize ?: config.mediaSizeLimit
-        return getMedia(uri, saveToCache, null, progress, maxSizeResolved)
+        return getMedia(uri, saveToCache, null, progress, maxSize)
     }
 
     override suspend fun getEncryptedMedia(
         encryptedFile: EncryptedFile,
         progress: MutableStateFlow<FileTransferProgress?>?,
         saveToCache: Boolean,
-        maxSize: FileSizeLimit?,
+        maxSize: Long?,
     ): Result<PlatformMedia> = kotlin.runCatching {
         val originalHash = encryptedFile.hashes["sha256"]
             ?: throw MediaValidationException(null, null)
-        val maxSizeResolved = maxSize ?: config.mediaSizeLimit
-        getMedia(encryptedFile.url, saveToCache, originalHash, progress, maxSizeResolved).getOrThrow()
+        getMedia(encryptedFile.url, saveToCache, originalHash, progress, maxSize).getOrThrow()
             .transformByteArrayFlow {
                 it.decryptAes256Ctr(
                     initialisationVector = encryptedFile.initialisationVector.decodeUnpaddedBase64Bytes(),
@@ -199,11 +194,10 @@ class MediaServiceImpl(
         animated: Boolean,
         progress: MutableStateFlow<FileTransferProgress?>?,
         saveToCache: Boolean,
-        maxSize: FileSizeLimit?,
+        maxSize: Long?,
     ): Result<PlatformMedia> = kotlin.runCatching {
         val thumbnailUrl = "$uri/${width}x$height/${api.json.encodeToJsonElement(method).jsonPrimitive.content}"
         val existingMedia = mediaStore.getMedia(thumbnailUrl)
-        val maxSizeResolved = maxSize ?: config.thumbnailSizeLimit
         if (existingMedia == null) {
             api.media.downloadThumbnailDependingOnServerVersion(
                 mxcUri = uri,
@@ -211,7 +205,7 @@ class MediaServiceImpl(
                 height = height,
                 method = method,
                 animated = animated,
-                maxSize = maxSizeResolved,
+                maxSize = maxSize,
                 progress = progress
             ) {
                 it.saveMedia(thumbnailUrl) { this }
@@ -230,7 +224,7 @@ class MediaServiceImpl(
         method: ThumbnailResizingMethod,
         animated: Boolean,
         progress: MutableStateFlow<FileTransferProgress?>? = null,
-        maxSize: FileSizeLimit,
+        maxSize: Long?,
         downloadHandler: suspend (Media) -> Unit
     ): Result<Unit> =
         if (serverDataStore.getServerData().versions.versions.contains(MATRIX_SPEC_1_11)) {