+++ trixnity-clientserverapi/trixnity-clientserverapi-client/src/commonMain/kotlin/de/connect2x/trixnity/clientserverapi/client/DownloadLimitExceededException.kt
@@ -1,20 +1,15 @@
 package de.connect2x.trixnity.clientserverapi.client
 
-import kotlin.math.ln
-import kotlin.math.pow
-import kotlin.math.roundToLong
+import kotlinx.coroutines.CopyableThrowable
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 
-class DownloadLimitExceededException(fileSizeLimit: Long) : IllegalStateException("File could not be downloaded because it would exceed the limit of ${fileSizeLimit.toHumanReadableSize()}")
-
-fun Long.toHumanReadableSize(): String {
-    if (this < 1024) return "$this B"
-    val exp = (ln(this.toDouble()) / ln(1024.0)).toInt()
-    val unit = arrayOf("KB", "MB", "GB", "TB", "PB")[exp - 1]
-    val value = this / 1024.0.pow(exp.toDouble())
-
-    val rounded = (value * 100).roundToLong()
-    val integer = rounded/100
-    val fraction = rounded % 100
-
-    return "$integer.$fraction $unit"
+@OptIn(ExperimentalCoroutinesApi::class)
+class DownloadLimitExceededException(
+    val maxSize: Long,
+    message: String? = null,
+    cause: Throwable? = null) : IllegalStateException("File could not be downloaded because it would exceed the limit of $maxSize bytes"),
+    CopyableThrowable<DownloadLimitExceededException> {
+    override fun createCopy(): DownloadLimitExceededException {
+        return DownloadLimitExceededException(maxSize, message, this)
+    }
 }